pyproject.toml
[project]
name = "aiautomatedarticleposting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.107.2",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
]
